requirements.txt
fastapi
uvicorn
pydantic
google-adk
google-generativeai
slowapi